[utils.py]: Moving the lock to utils.py.
