+++ utilities_common/cli.py
@@ -563,11 +563,11 @@ def json_dump(data):
         data, sort_keys=True, indent=2, ensure_ascii=False, default=json_serial
     )
 
-    
+
 def interface_is_untagged_member(db, interface_name):
-    """ Check if interface is already untagged member"""    
+    """ Check if interface is already untagged member"""
     vlan_member_table = db.get_table('VLAN_MEMBER')
-    
+
     for key,val in vlan_member_table.items():
         if(key[1] == interface_name):
             if (val['tagging_mode'] == 'untagged'):
